remove sshd.service since sshd package has included the service file.
This also fixes a problem with system-plugin-emulator and openssh. Currently
They conflict, because both provide sshd.service for systemd. But the service
file from this package is broken. It just runs /etc/init.d/ssh, but without
openssh package, this file doesn't exist.

This change re-enables working sshd when system-plugin-emulator is installed.

Change-Id: Ifd3671128549b18f7f3b92b92c9bb0f3346db65b
Signed-off-by: Rafal Krypa <r.krypa@samsung.com>