Fix mount configuration in %post
This package used to add three entries to /etc/fstab:

1. Bind mount of /opt/var to /var.
On Common it breaks the system, because /opt/var exists and contains only
few files. On IVI it's useless, because /opt/var doesn't exist.
On Mobile /opt/var is a hardlink to /opt, so it does nothing.
Removing this mount config.

2. Tmpfs mount for /tmp
This is useless on all profiles, because systemd already does that. And it
does it better, because it specifies more options, like Smack label.
Removing this mount config.

3. Swap on /dev/vdb
This works OK. for Common and Mobile. But IVI comes with no /etc/fstab, so
it doesn't get the swap.
Converting this mount config to systemd unit file.

Change-Id: Ic401f5ccb438fe2c17490247c6ef269de7820e13
Signed-off-by: Rafal Krypa <r.krypa@samsung.com>