wm_ready: add 30sec timeout on wm_read.service file
In general case, likely, wm_ready will be stopped within 5sec. But in newly
comming model, wm_ready can not be stopped by problem of xorg or
enlightenment. In this case, we have to ignore the initialization of
enlightenment. As a workaround(maybe no other choice), timeout was added
to wm_ready.service unit.

Change-Id: I08af4ab2ffb313801fb3c8d520223d6e4132ba80
Signed-off-by: Sooyoung Ha <yoosah.ha@samsung.com>